Search endpoint completed and larger reuseable functions made into helpers
